requirements.txt
streamlit==1.41.0
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.14
langchain_groq==0.2.2
pandas==2.2.3
python-dotenv==1.0.1
pydantic>=2.0.0
supabase>=2.3.0